requirements.txt
langchain==0.2.5
langchain-community==0.2.5
langchain-core==0.2.7
langchain-openai==0.1.8
langchain-text-splitters==0.2.1
openai==1.34.0
chromadb==0.5.0
python-dotenv==1.0.0
colorama==0.4.6
tiktoken==0.7.0
streamlit==1.24.0
streamlit-chat==0.1.1
pysqlite3-binary==0.5.3
